requirements.txt
python-telegram-bot==13.15
Pillow
gspread
oauth2client
pandas

